Do not put extractors options under CONF.extractor
We have defined CONF.extractor to configure the extractor that is going
to be used, but we are then defining in extractor.base CONF.extractor as
a section to hold the specific extractor options. Lets move these options
to extractor_nova and extrator_ceilometer and mark the deprecated group
name.