remove redundant help message
The help message for the extractor was printing the allowed values as
"dict_keys(['ceilometer', 'nova'])". The dict_keys is an ugly thing and
this is not needed, since we are defining the allowed values in the
configuration option and oslo is taking care of printing them in the
help message.